zyj 优化2018-6-1

+++ src/main/java/com/yun/common/Page.java
@@ -8,8 +8,8 @@ import java.util.List;
 @Data
 public class Page {
     private List<?> rows;
-    private int pageSize;  //页面显示的内容条数
-    private int total;		//总条数
-    private int pageNumber;   //
-    private int currentPage;   //当前页
+    private Long pageSize;  //页面显示的内容条数
+    private Long total;		//总条数
+    private Long pageNumber;   //
+    private Long currentPage;   //当前页
 }
